A lot of changes
I redid the way you navigate windows. Instead of destroying the window and creating a new root, I created a function that clears the widgets from the the window and adds new widgets  to make it look look a new window. I added a back button so that the user can go back between windows if they want. I added a save and load passwords function (two separate functions), by saving the passwords to a json file which makes it so that the user passwords passwords will be saved to a file and loaded again between program runs, so that their passwords would still be there. I also added a show/hide password button which give the user a sense of privacy when using the manager. I am looking to encrypt the view passwords button with a master password so that not everyone can access the users password willy nilly. The user will only have to memories the master password.